git-svn-id: https://svn.code.sf.net/p/asaac-openos/code/DeveloperBranches@34 1faf1f51-c62e-0410-9343-f9b3b0d0658f

+++ OpenOSLib/Source/Application/ApplicationMain.h
@@ -6,7 +6,7 @@ extern
 #ifdef __cplusplus
 "C"
 #endif
-long applicationMain( char argc, char** argv );
+int applicationMain( char argc, char** argv );
 
 
 extern
